frontend/notification_dispatcher/pyproject.toml
[project]
name = "notification_dispatcher"
version = "0.1.0"
description = "Notification dispatcher for the Telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python-ng>=2.2.3",
    "python-telegram-bot>=21.10",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.7.1",
]

[project.scripts]
notification_dispatcher = "notification_dispatcher.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
